Fixup to routing
