+++ lib/db.js
@@ -47,18 +47,6 @@ export const LAYOUT_SELECT = {
   }
 };
 
-// Fields that are server-generated or relational and must be stripped before
-// sending layout data to the DB (e.g. in SaveForm).
-export const LAYOUT_FILTER_KEYS = [
-  'user',
-  'createdAt',
-  'deletedAt',
-  'updatedAt',
-  'hearted',
-  '_count',
-  'userId',
-  'parentLayout'
-];
 
 export const layoutsQuery = {
   where: { deletedAt: null },
